python/requirements.txt
Flask
mysql-connector-python
pyjwt
requests
MagicMock
